requirements.txt
numpy
scipy
tqdm
tabulate
scikit-image
matplotlib
plyfile
tensorflow-gpu